Вынес тесты без параметризации в отдельные классы: - CatWithoutParamTest - LionExceptionTest
