+++ src/test/java/LionTest.java
@@ -66,16 +66,6 @@ public class LionTest {
         Assert.assertEquals(expected, actual);
     }
 
-    @Test
-    public void exceptionTest() throws Exception {
-        String sex = "Самцы";
-        String expectedMess = "Используйте допустимые значения пола животного - самец или самка";
-        Exception exception = assertThrows(Exception.class, () -> {
-            lion = new Lion(sex, feline);
-            });
-        String actualMess = exception.getMessage();
-        boolean actual = actualMess.contains(expectedMess);
-        Assert.assertEquals(true, actual);
-    }
+
 
 }
